Update dialog-polyfill.js to support open event
When the a dialog's `open` or `showModal` event is invoked, trigger an open event to correspond with the close event to comply w/ WHATWG expectations on how the dialog element should work. To not break the Principle of Least Surprise of the existing behavior of the Polyfill, the open event is only fired when `show` & `showModal` methods are explicitly used like how the `close` method currently behaves & differs from using the open attribute explicitly.