Support for explicit `onclose` handler
Some existing browser implementations of `HTMLDialogelement` handle the `onclose` property. Since `onclose` defines a single handler, it is extremely useful when reusing dialogs, as opposed to `addEventListener()`, where there is no ready or easy or clean way to clear all existing handlers. With `onclose` a dialog element can be used as a template and just have `onclose` assigned to whatever handler if and when necessary.