Fix findNearestDialog for child of slotted element
A focus event for an element which is a child of a slotted element
will be erroneously reported as not in the topmost dialog. This is
because for the slotted element we should follow assignedSlot rather
than the parentElement or parentNode to get the true hierarchy in
which the elmeent lives. This change adds handling for slotted
elements.